refactor(travel-expense): split 456-LOC component into 8 focused files
Behavior preserved 1:1. Structure mirrors app-shell/ convention.
