+++ components/trip-share-view.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { ArrowUpRight, Calendar, CheckCircle2, MapPin, Route, Users, Wallet } from 'lucide-react';
 import { Surface } from '@/components/ui';
-import { TravelExpenseTracker } from '@/components/travel-expense-tracker';
+import { TravelExpenseTracker } from '@/components/travel/travel-expense-tracker';
 import { type TravelTrip } from '@/data/site';
 
 const budgetStatusMeta = {
